cleaning, optimization class, normalisation of the dataset

+++ src/main/java/neuralnetwork/Layer.java
@@ -14,32 +14,32 @@ public class Layer {
         }
     }
 
-    public float[] forward(float[] inputs) {
-        float[] outputs = new float[neuronsLayer.length];
+    public double[] forward(double[] inputs) {
+        double[] outputs = new double[neuronsLayer.length];
 
 
         for (int i = 0; i < neuronsLayer.length; i++) {
             Neuron neuron = neuronsLayer[i];
 
-            float sum = 0;
+            double sum = 0;
             for (int j = 0; j < inputs.length; j++) {
                 sum += inputs[j] * neuron.weights[j];
             }
             sum += neuron.bias;
             neuron.lastInputs = inputs.clone();
 
-            float activated = neuron.activate(sum); //erase the linearity
+            double activated = neuron.activate(sum); //erase the linearity
             neuron.setValue(activated);
             outputs[i] = activated;
         }
         return outputs;
     }
-    public float[] backward(float[] errors, float learningRate) { // not understood completely
-        float[] prevErrors = new float[neuronsLayer[0].weights.length];
+    public double[] backward(double[] errors, double learningRate) { // not understood completely
+        double[] prevErrors = new double[neuronsLayer[0].weights.length];
 
         for (int i = 0; i < neuronsLayer.length; i++) {
             Neuron neuron = neuronsLayer[i];
-            float delta = errors[i] * neuron.derivative(neuron.getValue());
+            double delta = errors[i] * neuron.derivative(neuron.getValue());
             for(int j = 0; j < neuron.weights.length; j++) {
                 prevErrors[j] += delta * neuron.weights[j];
                 neuron.weights[j] -= learningRate * delta * neuron.lastInputs[j];
